% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\docType{data}
\name{myEnvModules}
\alias{myEnvModules}
\title{Global instance of RenvModule}
\format{
An object of class \code{RenvModule} of length 1.
}
\usage{
myEnvModules
}
\description{
Global instance of RenvModule
}
\keyword{datasets}
